+++ client/src/validation/schemas/videoUpdate.js
@@ -2,11 +2,11 @@ import { fields, createSchema, validate as _validate } from '..';
 
 const { requiredText, videoLink } = fields;
 
-const videoFeedback = createSchema({
+const videoUpdate = createSchema({
   link: videoLink,
   message: requiredText,
 });
 
-const validate = (data) => _validate(videoFeedback, data);
+const validate = (data) => _validate(videoUpdate, data);
 
 export default validate;
